Tests/LibWeb/Text/input/css/typed-om-set-empty-values.html

<!doctype html>
<html>
    <div id="foo"></div>
    <script src="../include.js"></script>
    <script>
        test(() => {
            foo.attributeStyleMap.set("width", CSS.px(10));
            println(`Initial style: ${foo.getAttribute("style")}`);

            try {
                foo.attributeStyleMap.set("width");
                println("FAIL: Empty set did not throw");
            } catch (e) {
                println(`Empty set threw ${e.name}`);
            }

            println(`After empty set: ${foo.getAttribute("style")}`);
        });
    </script>
</html>
