Tests/LibWeb/Text/expected/css/typed-om-set-empty-values.txt

Initial style: width: 10px;
Empty set threw TypeError
After empty set: width: 10px;
